Fix part of #10415: Refactoring improvements (#10952)
